assignment
Review of code.

The code for documenting test scores is not a bad code but there is no documenting at all from what i see, which isn't a problem for the code to run but for someone learning code or even just looking at the code it makes it harder to know what's going on. It doesn’t show examples of the purpose, purpose of the string and double, the purpose of what the grading means (line 23-39). Which means for someone who doesn't know code very well it makes it quite hard to understand. It also has a couple errors depending on what you put into the code example if you were to put the score mark of exactly 0, 40 or 100 which is test scores that can occur depending on how the test is marked then it will say “ O dear - you have put in a wrong test score” which I don’t believe was the intention when the code was being made as these are proper test scores. But overall there aren't that many errors in the code. But a couple things that could elevate the presented code would be the ability for a loop to line 17 if the individual did put the wrong test score as it would make it more convenient for them as well as on line 23- 25 they could have arranged it. If((dbl1 <= 40) && (dbl1 >= 0)) that would make the code more organized. I also used the >= and <= instead of just <,> as this will allow those numbers being 0 and 40 to be included as well as doing that with line 30’s < and > but line 34 has already done this at least.